Move `IpAddr`, `SocketAddr` and friends into core.
This is discussed in Rust RFC 2832, which is not accepted yet.

There seems to be no significant objection to including `IpAddr` and
friends. There is an objection to `SocketAddr` and friends:

> If the socketaddr types aren't identical across platforms, I don't
> think we want them in core.

Here is a possible interpretation:

core, and even the Rust language itself, already contains types with
platform-specific layouts. For example, `u64` has a platform-dependent
alignment, and that's visible even in otherwise fully encapsulating
types like `core::time::Duration`.

And core already has a `core::arch`, which is explicitly
architecture-dependent.

It seems the truly essential constraints of core are:

 - Nothing that does I/O. That goes in std.
 - No dynamic allocation. That goes in alloc.
 - No dependencies.

 - The public portable API feature set (eg. not `core::arch`) must be
   platform-independent. No unencapsulated `c_char` (signedness is
   platform-dependent) or `c_long` (range is platform-dependent).

   (Some types, like `c_schar`, could be debated, because one could
   argue that no serious platform would ever make `c_schar` different
   from `i8` ever again, but it isn't urgent to settle this.)

 - There must be fallback code so that core can be ported without any
   OS-specific work.

So I propose the path is: add the socket types to core, with
encapsulated OS-specific internal representations. The public-facing
feature set will be platform-independent. And add a `core::os` module,
similar to `core::arch` but for OS-specific features, which defines
`Ext` extension traits, similar to `std::os::*`, which expose the
platform-specific internal representations for FFI use cases.

A platform with no `core::os` support will still be able to use `core`
without any porting work, they just won't be able to access the
internal representations, which isn't a regression.